Updated readme, packages and statusMessage deleting

+++ events/player.events/queueDelete.js
@@ -6,7 +6,9 @@ module.exports = {
   async execute(queue) {
     logInfo(`[${queue.guild.name}] Queue deleted`);
     try {
-      await queue.metadata.statusMessage.delete();
+      await new Promise((resolve) => setTimeout(resolve, 1000));
+      if (queue.metadata.statusMessage)
+        await queue.metadata.statusMessage.delete();
     } catch (error) {
       logInfo(`Unable to delete status message`, error.message);
     }
